Support multiple monitors

+++ library/include/borealis/platforms/glfw/glfw_platform.hpp
@@ -32,7 +32,7 @@ class GLFWPlatform : public DesktopPlatform
     ~GLFWPlatform();
 
     std::string getName() override;
-    void createWindow(std::string windowTitle, uint32_t windowWidth, uint32_t windowHeight, int windowXPos, int windowYPos) override;
+    void createWindow(std::string windowTitle, uint32_t windowWidth, uint32_t windowHeight, float windowXPos, float windowYPos) override;
 
     virtual void restoreWindow() override;
     virtual void setWindowSize(uint32_t windowWidth, uint32_t windowHeight) override;
